workaround a race in VNC API
handler of project gets makes sure that an application_policy_set
instance for the uuid exists and attempts to create it if there is
no one. this occurs before read of the project entity itself thus
it is not clear if the main entity exists to avoid the auxilliary
activity. then the create fails with NoIdError and the get operation
reports 500 instead of 404. this confuses tests, they produce
false negative results.
fortunatelly, 500 error response contains info about reason of the
failure, thus we can interpret it and not fail.

Change-Id: I2193c5b9f5c2fb2e1bf0c39a450283d781b152bc
Signed-off-by: sgalas@progmaticlab.com