DPDK Vrouter agent restart command sequences/script changes for RHOSP16 environment.
The vrouter agent needs to be stopped before restarting dpdk agent
and then started after dpdk-agent is up. Doing this using podman
start/stop does not help because systemd brings up the vrouter
agent as soon as podman stops it. This fix is to handle this
specific scenario faced in rhosp-16.

Issue-ID: CEM-27587
Signed-off-by: ritam <ritam@juniper.net>
Change-Id: I1f358c2645293e6ea52b37442da054cbf7c5e43d